delay time must be float

+++ NSGIF/NSGIF.m
@@ -81,7 +81,7 @@ typedef NS_ENUM(NSInteger, GIFSize) {
 
 }
 
-+ (void)createGIFfromURL:(NSURL*)videoURL withFrameCount:(int)frameCount delayTime:(int)delayTime loopCount:(int)loopCount completion:(void(^)(NSURL *GifURL))completionBlock {
++ (void)createGIFfromURL:(NSURL*)videoURL withFrameCount:(int)frameCount delayTime:(float)delayTime loopCount:(int)loopCount completion:(void(^)(NSURL *GifURL))completionBlock {
     
     // Convert the video at the given URL to a GIF, and return the GIF's URL if it was created.
     // The frames are spaced evenly over the video, and each has the same duration.
@@ -224,7 +224,7 @@ CGImageRef ImageWithScale(CGImageRef imageRef, float scale) {
              };
 }
 
-+ (NSDictionary *)framePropertiesWithDelayTime:(int)delayTime {
++ (NSDictionary *)framePropertiesWithDelayTime:(float)delayTime {
 
     return @{(NSString *)kCGImagePropertyGIFDictionary:
                 @{(NSString *)kCGImagePropertyGIFDelayTime: @(delayTime)},
